sampleLinkedList and code refactoring
